refactor: split handle-rewiring tests out, simplify a guard clause
